Add dependency-budgetting system
Files loaded as dependencies of other files now have to fit inside the
'budget' of that file, set by the dependencyBudget server option
(counted in number of expressions, defaults to 20k) to be analyzed.

Dependencies are added in a breadth-first fashion until the parent's
budget is exhausted. Dependencies of dependencies count towards the
original importer's budget.

This is designed to address the problem of accidentally pulling in
giant libraries through the node plugin, and killing the Tern server
this way.

Issue marijnh/tern_for_sublime#32