Deleting CR to adjust the text position to editor
Issue #467
If the source code is encoded in DOS-like line separator,
Emacs reports incorrect type definition or throw errors.
It is caused by the difference of handling EOL between Emacs and
node.js, Emacs treats CRLF as one character, but node.js does as 2
characters.

This patch deletes the CR character from the source codes
which are loaded on the memory and adjust to Emacs handling.